+++ tests/dummy/app/templates/components/media-detail.hbs
@@ -1,0 +1,34 @@
+<Boxel @tag="section" @class="item">
+  {{#each this.detailSections as |s|}}
+    <section class="item__section">
+      <h3 class="item__section-title">{{s.title}}</h3>
+      {{#each s.content as |field|}}
+        <div class="item__field">
+          <h4 class="item__field-title">{{humanize field.title}}</h4>
+          {{#if field.value.id}}
+            <Card @variant="small" @model={{field.value}} />
+          {{else if (is-array field.value)}}
+            <div class="item__field-value">
+              {{#each field.value as |val|}}
+                {{#if val.id}}
+                  <Card @variant="small" @model={{val}} />
+                {{else if (is-array val.value)}}
+                  {{#if val.title}}
+                    <div class="item__field-value-title">{{val.title}}</div>
+                  {{/if}}
+                  {{#each val.value as |vv|}}
+                    <div class="item__field-value-val">{{vv}}</div>
+                  {{/each}}
+                {{else}}
+                  <div class="item__field-value">{{val}}</div>
+                {{/if}}
+              {{/each}}
+            </div>
+          {{else}}
+            <div class="item__field-value">{{field.value}}</div>
+          {{/if}}
+        </div>
+      {{/each}}
+    </section>
+  {{/each}}
+</Boxel>
