+++ tests/dummy/app/templates/media-registry/item/edit.hbs
@@ -1,0 +1,5 @@
+<BoxelModal @onClose={{this.transitionToPrevious}}>
+  <div class="edit-modal">
+    <MediaDetail @model={{@model}} />
+  </div>
+</BoxelModal>
